[feat] 프로젝트 삭제, PM 가져오기

+++ src/main/java/kakao99/backend/project/service/ProjectService.java
@@ -9,6 +9,7 @@ import kakao99.backend.group.dto.GroupNameDTO;
 import kakao99.backend.group.repository.GroupRepository;
 import kakao99.backend.project.dto.ProjectDTO;
 import kakao99.backend.project.dto.ProjectModifyDTO;
+import kakao99.backend.project.dto.ProjectPMDTO;
 import kakao99.backend.project.repository.MemberProjectRepository;
 import kakao99.backend.project.repository.ProjectRepository;
 import kakao99.backend.utils.ResponseMessage;
@@ -19,9 +20,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import java.util.Date;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.concurrent.ExecutionException;
+
+import static io.lettuce.core.pubsub.PubSubOutput.Type.message;
 
 @Service
 @RequiredArgsConstructor
@@ -71,27 +73,66 @@ public class ProjectService {
         catch(Exception e){
             ResponseMessage message = new ResponseMessage(200, "프로젝트 수정 실패");
 
-            return new ResponseEntity(message, HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
         ResponseMessage message = new ResponseMessage(200, "프로젝트 수정 성공");
-        return new ResponseEntity(message, HttpStatus.OK);
+        return new ResponseEntity<>(message, HttpStatus.OK);
     }
     @Transactional
-    public ResponseEntity<?> removeProject(ProjectModifyDTO projectModifyDTO){
-        try {
-            Optional<Project> optionalProject = projectRepository.findById(projectModifyDTO.getId());
-            Project project = optionalProject.get();
+    public ResponseEntity<?> removeProject(ProjectModifyDTO projectModifyDTO, Long memberId){
+        Optional<Project> optionalProject = projectRepository.findById(projectModifyDTO.getId());
+        System.out.println(projectModifyDTO.getId());
+        if(optionalProject.isEmpty()) {
+            ResponseMessage message = new ResponseMessage(500, "id로 삭제할 프로젝트 확인 실패");
+
+            return new ResponseEntity<>(message, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        Project project = optionalProject.get();
+        Optional<String> opRole = memberProjectRepository.findRole(project.getId(), memberId);
+        if(opRole.isEmpty()){
+            ResponseMessage message = new ResponseMessage(500, "id로 찾은 role 확인 실패");
+
+            return new ResponseEntity<>(message, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        String role = opRole.get();
+        System.out.println(role);
+        if(role.equals("PM")) {
             memberProjectRepository.deleteMemberProjectByProjectId(projectModifyDTO.getId());
             project.deleteProject();
+            ResponseMessage message = new ResponseMessage(200, "프로젝트 삭제 성공");
+            return new ResponseEntity<>(message, HttpStatus.OK);
+        }else{
+            ResponseMessage message = new ResponseMessage(200, "프로젝트 삭제 권한이 없습니다.");
+            return new ResponseEntity<>(message, HttpStatus.OK);
         }
-        catch(Exception e){
-            ResponseMessage message = new ResponseMessage(500, "프로젝트 삭제 실패");
-
-            return new ResponseEntity(message, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-
-        ResponseMessage message = new ResponseMessage(200, "프로젝트 삭제 성공");
-        return new ResponseEntity(message, HttpStatus.OK);
     }
+
+    @Transactional
+    public List<ProjectPMDTO> makeProjectPMDTOS(List<Project> projects) throws Exception {
+        List<ProjectPMDTO> projectPMDTOS = new ArrayList<>();
+        for (Project project : projects) {
+            System.out.println(project.getId());
+            Optional<Member> optionalMember = memberProjectRepository.findPMByProjectId(project.getId(), "PM");
+            if(optionalMember.isEmpty()){
+                throw new Exception("PM이 없음");
+            }
+            Member pmMember = optionalMember.get();
+            ProjectPMDTO projectPMDTO = ProjectPMDTO.builder()
+                    .PMId(pmMember.getId())
+                    .PMName(pmMember.getNickname())
+                    .id(project.getId())
+                    .name(project.getName())
+                    .description(project.getDescription())
+                    .createdAt(project.getCreatedAt())
+                    .groupId(project.getGroup().getId())
+                    .isActive(project.getIsActive())
+                    .updatedAt(project.getUpdatedAt())
+                    .build();
+            projectPMDTOS.add(projectPMDTO);
+        }
+
+        return projectPMDTOS;
+    }
+
 }
